% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.r
\name{doit}
\alias{doit}
\title{doit}
\usage{
doit(meta)
}
\arguments{
\item{meta}{the name of the meta-analysis (e.g "GxSES")}
}
\value{
-
}
\description{
doit does a meta-analysis
}
\details{

}
\examples{
\dontrun{
doit("GxSES")
}
}
\seealso{
Other Reporting Functions: \code{\link{meta.now}}
}
