% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.r
\name{load_bib}
\alias{load_bib}
\title{Read in the bib file containing references to be cited}
\usage{
load_bib(meta)
}
\arguments{
\item{meta}{name of the meta-analytic "question"}
}
\value{
- the effects data.frame
}
\description{
You provide the name of the meta analysis (say "GxSES") which is expanded to "bib_GxSES.bib"
and opened (the actual file resides in the package's data folder).
}
\details{

}
\examples{
bib = load_bib(meta = "GxSES") # loaded as bib
}
\seealso{
Other Data Functions: \code{\link{load_meta}},
  \code{\link{meta.now}}
}
