% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.r
\name{load_meta}
\alias{load_meta}
\title{Read in the xlsx file containing effects to be meta-analysed}
\usage{
load_meta(meta, verbose = getOption("verbose"))
}
\arguments{
\item{meta}{name of the meta-analytic "question"}
}
\value{
- the effects data.frame
}
\description{
You provide the name of the meta analysis (say "GxSES") which is expanded to "data_GxSES.xlsx"
and opened (the actual file resides in the package's data folder).
}
\details{

}
\examples{
load_meta(meta = "GxSES") # loaded as df
}
\seealso{
Other Data Functions: \code{\link{load_bib}},
  \code{\link{meta.now}}
}
