% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\docType{package}
\name{meta.now}
\alias{meta.now}
\alias{meta.now-package}
\title{Do a meta-analysis: now!}
\description{
meta.now helps keep meta-analyses up-to-date, with functions to produce the graphs and text of a meta-analysis,
and support via git for users to push updates as new data emerge
}
\details{
If you want the bleeding-edge version:

devtools::install_github("tbates/meta.now")
}
\examples{

}
\seealso{
Other Reporting Functions: \code{\link{doit}}

Other Data Functions: \code{\link{load_bib}},
  \code{\link{load_meta}}

Other Miscellaneous Utility Functions: \code{\link{meta_make}}
}
