% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{meta_make}
\alias{meta_make}
\title{make meta.now using devtools}
\usage{
meta_make(what = c("install", "examples", "check", "win", "rhub", "release"),
  pkg = "~/bin/meta.now", check = TRUE)
}
\arguments{
\item{what}{whether to "install", "release" to CRAN, check on "win", "check", or "examples"))}

\item{pkg}{the local path to your package. Defaults to my path to meta.now}

\item{check}{Whether to run check on the package before release (default = TRUE)}
}
\value{
-
}
\description{
Easily  run devtools "install", "release", "win", or "examples".
}
\examples{
\dontrun{
umx_make(what = "install"))  # just installs the package
umx_make(what = "examples")) # run the examples
umx_make(what = "check"))    # run R CMD check
umx_make(what = "win"))      # check on win-builder
umx_make(what = "release"))  # release to CRAN
}
}
\seealso{
Other Miscellaneous Utility Functions: \code{\link{meta.now}}
}
