% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{load_meta}
\alias{load_meta}
\title{Read in the xlsx file containing effects to be meta-analysed}
\usage{
load_meta(meta, type = "xlsx", verbose = getOption("verbose"))
}
\arguments{
\item{meta}{name of the meta-analytic "question"}
}
\value{
- the effects data.frame
}
\description{
You provide the name of the meta analysis (e.g. "GxSES") which is expanded to the file name containing 
all the preior estiamtes, e.g. "data_GxSES.xlsx".
}
\details{
(the actual file resides in the package's data folder).
}
\examples{
df = load_meta(meta = "GxSES") # loaded as df
}
\seealso{
Other Data Functions: \code{\link{load_bib}},
  \code{\link{meta.now}}
}
